+++ webapp/src/modules/translation/locales/es.json
@@ -361,6 +361,12 @@
       "less_than_price": "Menos de <mana></mana> {price}",
       "price_between": "<mana></mana> {minPrice} - <mana></mana> {maxPrice}"
     },
+    "estate_size": {
+      "all_sizes": "Todos los tamaños",
+      "more_than_size": "Más de {size} parcelas",
+      "less_than_size": "Menos de {size} parcelas",
+      "size_between": "{min} a {max} parcelas"
+    },
     "collection": {
       "title": "Colección",
       "all_items": "Todas las colecciones",
